+++ cli/package.json
@@ -5,7 +5,8 @@
 	"type": "module",
 	"main": "dist/mod.js",
 	"scripts": {
-		"build": "tsc --build"
+		"build": "tsc --build",
+		"cli": "node dist/mod.js"
 	},
 	"author": "wolfpup",
 	"license": "BSD-3-Clause"
