+++ package.json
@@ -1,13 +1,16 @@
 {
     "workspaces": [
         "core",
-        "cli"
+        "cli",
+        "tests"
     ],
     "scripts": {
         "build": "npm run --workspaces build",
-        "fmt": "npx prettier --write ./"
+        "fmt": "npx prettier --write ./",
+        "test": "node ./cli/dist/mod.js --file ../../tests/dist/mod.js"
     },
     "devDependencies": {
+        "@types/node": "^22.5.5",
         "prettier": "^3.3.1",
         "typescript": "^5.4.5"
     }
