+++ tests/package.json
@@ -1,0 +1,12 @@
+{
+	"name": "tests",
+	"author": "wolfpup",
+	"version": "0.1.0",
+	"description": "",
+	"license": "BSD-3-Clause",
+	"type": "module",
+	"main": "dist/mod.js",
+	"scripts": {
+		"build": "tsc --build"
+	}
+}
