+++ tests/src/mod.ts
@@ -1,0 +1,17 @@
+function testTheStuff() {
+	return ["this test failed!"];
+}
+
+let collection = {
+	title: "Demo_test",
+	runAsynchronously: false,
+	tests: [
+		testTheStuff,
+	]
+}
+
+let collections = [
+	collection
+]
+
+export { collections }
